chargeSim: Enable basic drawing to screen
Makefile somewhat hackily makes chargeSim for linux and also copies relevant
files to a "win" symlink to allow for building with mingw64. Winmake is the
mingw64 makefile while Makefile is the linux makefile.
chargeSim contains the main() logic, but not much else.
Added a Screen object to gameScreen which contains SDL instances.
Added some temporary test code to make sure that we actually get rendering.
draw() in it's current state is experimental at best.